Make sure you got a slot before calling that method son
git-svn-id: http://svn.apostrophenow.org/plugins/apostropheExtraSlotsPlugin/trunk@4398 c89fab08-2f9b-42bf-8176-bd3ad6bee547

+++ lib/model/doctrine/PluginaReusableSlotTable.class.php
@@ -61,9 +61,12 @@ class PluginaReusableSlotTable extends Doctrine_Table
   static public function purgeOrphanByLabel($label)
   {
     $slot = Doctrine::getTable('aReusableSlot')->findOneByLabel($label);
-    if (!$slot->getReusedSlot())
+    if ($slot)
     {
-      $slot->delete();
+      if (!$slot->getReusedSlot())
+      {
+        $slot->delete();
+      }
     }
   }  
 }
